// src/app/models/userData.model.ts
export interface UserData {
  uid: string;
  email: string;
  country: string;
  date: number;
  displayName: string;
  photoURL: string;
  phone: number;
  role: string;
  status: number;
  verified: number;
  requestedVerification: number
  address: string;
  username: string;
  currency: any;
  phoneCode: any;
}